requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.26.0
pandas>=2.0.0
tqdm>=4.0.0
python-dotenv>=1.0.0